+++ artifact/scripts/Base/basic.lua
@@ -1,3 +1,8 @@
+require(".\\Base\\scope")
+basicInit=true
+if scopeInit then
+	print("scope script loaded")
+end
 local api = uevr.api
 local vr = uevr.params.vr
