+++ artifact/scripts/Base/scope.lua
@@ -1,5 +1,8 @@
-require("Trackers")
-
+require(".\\Base\\HolsterScript")
+scopeInit=true
+if HolsterInit then
+	print("HolsterScript loaded")
+end
 local api = uevr.api
 local vr = uevr.params.vr
