+++ artifact/scripts/Entry.lua
@@ -1,0 +1,8 @@
+require(".\\Base\\basic")
+
+
+if basicInit then
+	print("basic lua loaded")
+	print("Luas loaded")
+end
+
